Rework SQL Server tests to use BaseConnectorTest
